move `qml.device` to a dedicated file
